Add option for adding function signatures to the binaries.
The function signatures aren't a standard part of how other systems
generate their binaries, as they tend to have support for run
time linking, and thus have symbols in the file as linkage
information. However, as I'm not currently doing that, and am instead
trying to give myself a way to work this into the existing pattern
of ARM32 execution, I've added the function signatures to the
binary. It's optional, and does require a post-process stage to
make it happen, but it seems to be working ok.