Updated cog to animate properly.
The cog now animates properly, with bank flipping to make it smooth
and we use a short delay to make it display at a sensible speed.

The SWIs have been updated slightly to add the calls that I need, and
the cog data has been reduced as I realised that obviously I don't need
to duplicate the path data for each frame - that's what the matrices
are for!