Add a more complete C library.
From the other work in compiling applications for RISC OS we have
more of a C library now. This library just brings in the functions
that we have needed. In particular, the malloc function is pretty
poor - just a static heap, which we allocate from and never free.

SWIs are now implemented better, with _kernel_swi and _swix provided,
although the latter doesn't handle flags yet.

The header is also updated to include the entry points for 3 expected
initialisations - decompress (should be handled by the Kernel, if we
use at all), zero initialisation and then the startup of the application.

Application startup now sets up the argv string as well.