Make the heap allocation use application space; reduce linkages.
The code has been reorganised so that the heap is now the application
space, up to the stack limit (which we don't enforce), and some files
split up so that they don't get linked in if we don't need it. In
particular, we don't get malloc or most of the string functions if
they are not required.