Add sin, cos, some kernel calls and sysvars; plus more tests.
The tests for the clib which just check that we're sane now include
the crash code, and the hello world using printf, which also tests
that main parsing and the _clib_version message.

sin and cos are now implemented and we have some stub module entries
that should let us start up.

System variables support with getenv and _kernel_set/getenv are present.
We have the strspn and strcspn calls too.