Added FILE * cleanup on exit, and base linkage reduction; tentative mods.
We now clean up the FILEs that were opened by the system when the
application shuts down. These FILEs should be closed and their workspace
freed when the application shuts down. The 'free' function is marked as
weak, so that we only use it if there are actual allocations present
in the code - ie if we never open files, we'll never need to pull in
the malloc code.

System variables have been handled the same way with a free function
that's only called if there was any memory allocation.

The size of the signatured functions has been reduced to 5 words (20
bytes) which means that longer strings may not be given function
signatures if we add them. Warnigns haev ben added to the riscos_symbols
script to alert about this.

The exit handling has been reworked so that we don't tie it to the main
handling, so that it can be used when we have modules linked with us.
Stub _clib_initialisemodule and _clib_finalisemodule (which just calls
the regular finalise) are now persent.