Make the modules build with headers.
Trying to move towards a structure for this test code that is
similar to that for the RISC OS Pyromaniac module tests - these
will eventually be moved to the source as the tests.

This change introduces the sources for SWI numbers and moves the
simple print macros to a header.