Added heap implementation using OS_Heap and RMA.
We can now allocate memory through OS_Heap when you're an application
or through the RMA if a module is in use. This isn't quite how it was
before, but we're closer to the same sort of state. The SWIs for the
module and the heap calls are separated so that they don't link at
other times.

The module initialisation test code has been fixed so that it returns
errors properly now as well.