Replace the heap implementatio with O(1)heap.
The OS_Heap implementation is small in the compiled code, but the
speed may impact the system badly. The O(1)heap I've added in here
works much faster and is still quite small. The heap system has
been restructured so that we can select between the heaps. For
simplicity these are C library compile time, not compile time of
the target application (or runtime selection). In the future it
could be improved.