Refactor the heap functions into a dispatcher structure.
The heap management was ok, but having it referenced by function names
meant that you had to rebuild the library in order to change the heap
type. Now, we can use the different heap implementations independantly
without having to worry about the linkage - just change the
__heap_implementation variable to something that contains the structure
and the implementation will be modified.