Improve the printf formatting with more aligned fields and more formats.
We now support more field alignments, and more formats for the printf
formats. The byte, halfword, size_t and ptrdiff_t format modifiers are
now supported and print suitable values. The left and right alignment
on strings, hex and decimal printed values should now be honoured.
The alternate forms of the string output should be honoured (although
they might not be quite right).

We now have some quite simple test code that tries to exercise some
of the printf formats to check that we have sensible output.