Add abort, atoi, toupper/lower, remove, rename, tmpnam; fix some.
The abort function is now implemented and should abort the program
with a suitable backtrace..

atoi uses the strtol function.

toupper/tolower as ASCII-focused.

remove, rename and tmpnam are rudimentary implementations of the
standard FS functions.

clock will now try to base itself off the start of the program, rather
than the start of the OS.