Add a slightly more optimised memset implementation.
In VideoTTX the memsets at the start of the initialisation were
being quite long and when traced were taking a long time. The change
here applies the 64bit data writes in sections so that we can do
more stores. This still isn't as fast as we could go - we could use
STP as well, but I'm not quite sure how to trigger that in GCC.

For now, this gives a lot less trace output, and so we're much more
efficient.