Rework the way in which everything is linked so for fPIC.
We need to support fPIC so that modules can work. In order for modules
to work, they need to be able to be linked in a fPIC form. This means
that we need runtime relocations. To have the runtime relocations we
need to create relocation tables. Renaming some of the files to be
more sensible (the riscos64-* files) and changing some of the symbols
to be more sensible (in the linker scripts) also came out of this.