Improved debugging and fix for missing signatures in addsignatures.
The addsignatures utility would fail to add the signatures to the
code if there were values that seemed to be lower than the address
base - which would always happen for the symbols for the data
section size, it seems. We now explicitly look for the start of the
.text section.

Debugging improved and added explicit reading of the file as binary.