Fix for the generation of the relocation table breaking the table itself.
The relocation table was being generated into the .rodata section.
However, the references that the relocation table was using were
themselves within the .rodata section. So adding the extra table was
purturbing the addresses and causing us to have invalid relocations.

This has now been fixed by creating a separate dedicated section for
the relocation data during the implicit linkage.