Simplify and improve I/O functions.
Rather than checking for NULL in every I/O function, we now bundle this
check into the CHECK_MAGIC which is performed everywhere. This should
make it clearer in the code. Where there are issues, the errno value
has been updated as well, so that it contains the right value when we
return failures due to invalid permissions.