Move the line reading function into the I/O dispatcher.
Because line reading for console is different to the file - it can
be edited - we want to have separate code in this. The classic
RISC OS C library had this in a separate function and handled it
itself. We just call OS_ReadLine. This does, however, prevent the
EOF case.