Add support for opening memory streams as FILEs.
The `open_memstream` function is now implemented. This is a function
that is defined by POSIX, which creates a dynamic buffer to write the
output to. This content is accessible after the file is closed.

The functionality is complete now, and seems to be working for the
chibbic compiler - I can compile hellow world and then assemble the
resulting code, and link, and it works on RISC OS 64.